Put the types for notifications store in the central types location
