+++ packages/client/src/types/store.ts
@@ -24,7 +24,6 @@ import {
 import { Action, NotifVariant } from "@/enums/store";
 import { CustomerRoute } from "@/enums/routes";
 import { DocumentData } from "firebase/firestore";
-import { NotificationsState } from "@/store/reducers/notificationsReducer";
 
 // #region app
 /**
@@ -263,6 +262,40 @@ export type FirestoreState = {
   listeners: { [index in CollectionSubscription]?: FirestoreListener };
 };
 // #endregion firestore
+
+// #region notifications
+/**
+ * A notification to be enqueued in the store and shown to the UI
+ */
+export interface NotificationInterface {
+  /** A uuid string generated to identify the notification */
+  key: string;
+  /** Content */
+  message: string;
+  /** "success" / "error" */
+  variant: NotifVariant;
+}
+
+/**
+ * Notification porton of the store is in charge of enqueueing notifications and displaying them
+ * in an orderly manner.
+ */
+export interface NotificationsState {
+  /**
+   * Queue houses all notifications showing or to be shown, where
+   * the first one (queue[0]) is the one currently shown.
+   */
+  queue: NotificationInterface[];
+  /**
+   * This flag is `true` in special cases where there's only one notification in store,
+   * but has been shown for long enough that it can be replaced with the new one.
+   *
+   * If `false`, any new notification being added will simply be enqueued at the end of the queue
+   */
+  canEvict: boolean;
+}
+
+// #endregion notifications
 
 // #region thunks
 type GetState = () => LocalStore;
